+++ packages/website/src/app/shared/markdown.ts
@@ -14,7 +14,8 @@ export function markedOptionsFactory() {
 
   renderer.heading = (text, level, raw) => {
     const slug = slugger.slug(raw);
-    const anchorLink = `<a class="heading-anchor" href="javascript:void(0)" onclick="window.location.hash = '${slug}'" aria-hidden="true">#</a>`;
+    text = text.replace(/"/g, '&quot;');
+    const anchorLink = `<a class="heading-anchor" href="javascript:void(0)" onclick="window.location.hash = '${slug}'" aria-hidden="true" aria-label="${text}">#</a>`;
     return `<h${level} id="${slug}" class="heading">${text} ${anchorLink}</h${level}>`;
   };
 
@@ -29,6 +30,13 @@ export function markedOptionsFactory() {
   };
 
   return {
+    hooks: {
+      preprocess: (markdown: string) => {
+        slugger.reset();
+        return markdown;
+      },
+      postprocess: (html: string) => html
+    },
     renderer: renderer
   };
 }