+++ packages/website/src/app/workshop/pagination.component.ts
@@ -2,6 +2,7 @@ import { Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Workshop } from './workshop';
 import { setHash, setQueryParams } from '../router';
+import { resetFocus } from '../shared/focus';
 
 @Component({
   selector: 'app-pagination',
@@ -9,12 +10,12 @@ import { setHash, setQueryParams } from '../router';
   imports: [CommonModule],
   template: `
     <nav class="pagination" aria-label="Page navigation">
-      <button [hidden]="!hasPrevious()" (click)="previous()" class="previous" aria-label="Go to previous page">
-        <div class="label">Previous</div>
+      <button [hidden]="!hasPrevious()" (click)="previous()" class="previous">
+        <div class="label" aria-label="Go to previous page:">Previous</div>
         <div class="title">{{ getTitle(-1) }}</div>
       </button>
-      <button [hidden]="!hasNext()" (click)="next()" class="next" aria-label="Go to next page">
-        <div class="label">Next</div>
+      <button [hidden]="!hasNext()" (click)="next()" class="next">
+        <div class="label" aria-label="Go to next page:">Next</div>
         <div class="title">{{ getTitle(1) }}</div>
       </button>
     </nav>
@@ -36,7 +37,7 @@ import { setHash, setQueryParams } from '../router';
         text-align: left;
 
         > .title::before {
-          content: '←';
+          content: '←' / '';
           margin-right: var(--space-xxs);
         }
       }
@@ -46,7 +47,7 @@ import { setHash, setQueryParams } from '../router';
         text-align: right;
 
         > .title::after {
-          content: '→';
+          content: '→' / '';
           margin-left: var(--space-xxs);
         }
       }
@@ -98,6 +99,7 @@ export class PaginationComponent {
     if (this.hasPrevious()) {
       setQueryParams({ step: this.workshop!.step - 1 });
       setHash();
+      resetFocus();
     }
   }
 
@@ -105,6 +107,7 @@ export class PaginationComponent {
     if (this.hasNext()) {
       setQueryParams({ step: this.workshop!.step + 1 });
       setHash();
+      resetFocus();
     }
   }
 }