+++ packages/website/src/app/workshop/workshop.component.ts
@@ -30,8 +30,12 @@ import { getRepoPath } from '../shared/loader';
   template: `
     <div (click)="sidebar.toggleOpen(false)" class="full-viewport">
       <app-header [title]="workshop?.shortTitle || 'Workshop'" [sidebar]="sidebar"></app-header>
-      <div class="content">
-        <app-sidebar #sidebar="sidebar" [links]="menuLinks"></app-sidebar>
+      <main class="content">
+        <app-sidebar
+          #sidebar="sidebar"
+          [links]="menuLinks"
+          [numbering]="workshop?.meta?.navigation_numbering"
+        ></app-sidebar>
         <app-loader
           id="workshop"
           [loading]="loading"
@@ -54,7 +58,7 @@ import { getRepoPath } from '../shared/loader';
           </div>
           <app-footer></app-footer>
         </app-loader>
-      </div>
+      </main>
       <ng-template #noWorkshop>
         <p class="container" *ngIf="!loading">Could not load workshop :(</p>
       </ng-template>